Use NestMiddleware to register token handlers
